Support reading/writing multiple keys.
This is a first step towards addressing #13, as we currently store session keys and stored keys with the same prefix, and distinguish them only based on which type of storage they are in.  With #13, we will need to have different keys with different purposes concurrently stored in the same type of storage.